more and more

+++ app/displays/hd44780_i2c.py
@@ -104,8 +104,8 @@ class hd44780_i2c():
 
 		self.rows = rows
 		self.cols = cols
-		self.rows_char = rows/8
-		self.cols_char = cols/5
+		self.rows_char = int(rows/8)
+		self.cols_char = int(cols/5)
 		self.curposition = (0,0)
 
 		self.enable_duration = enable_duration
